+++ spring-cloud-starter-tencent-polaris-ratelimit/src/main/java/com/tencent/cloud/polaris/ratelimit/config/PolarisRateLimitConfiguration.java
@@ -19,6 +19,7 @@
 package com.tencent.cloud.polaris.ratelimit.config;
 
 import com.tencent.cloud.polaris.context.ConditionalOnPolarisEnabled;
+import com.tencent.cloud.polaris.context.PolarisContextAutoConfiguration;
 import com.tencent.cloud.polaris.context.ServiceRuleManager;
 import com.tencent.cloud.polaris.ratelimit.RateLimitRuleLabelResolver;
 import com.tencent.cloud.polaris.ratelimit.constant.RateLimitConstant;
@@ -30,6 +31,7 @@ import com.tencent.polaris.client.api.SDKContext;
 import com.tencent.polaris.ratelimit.api.core.LimitAPI;
 import com.tencent.polaris.ratelimit.factory.LimitAPIFactory;
 
+import org.springframework.boot.autoconfigure.AutoConfigureAfter;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnWebApplication;
@@ -45,18 +47,15 @@ import static javax.servlet.DispatcherType.INCLUDE;
 import static javax.servlet.DispatcherType.REQUEST;
 
 /**
+ * Configuration of rate limit.
+ *
  * @author Haotian Zhang
  */
 @Configuration
 @ConditionalOnPolarisEnabled
-@ConditionalOnProperty(name = "spring.cloud.polaris.ratelimit.enabled",
-		matchIfMissing = true)
-public class RateLimitConfiguration {
-
-	@Bean
-	public PolarisRateLimitProperties polarisRateLimitProperties() {
-		return new PolarisRateLimitProperties();
-	}
+@AutoConfigureAfter(PolarisContextAutoConfiguration.class)
+@ConditionalOnProperty(name = "spring.cloud.polaris.ratelimit.enabled", matchIfMissing = true)
+public class PolarisRateLimitConfiguration {
 
 	@Bean
 	@ConditionalOnMissingBean