+++ CHANGELOG.md
@@ -1,7 +1,6 @@
 # Change Log
 ---
 
-
 - [Feature: Support parse ratelimit rule expression labels.](https://github.com/Tencent/spring-cloud-tencent/pull/183)
 - [Feature: Router support request label.](https://github.com/Tencent/spring-cloud-tencent/pull/165)
 - [Feature: Support router expression label](https://github.com/Tencent/spring-cloud-tencent/pull/190)
@@ -9,3 +8,5 @@
 - [Feature: Support metadata router.](https://github.com/Tencent/spring-cloud-tencent/pull/191)
 - [Feature: Misc optimize metadata router.](https://github.com/Tencent/spring-cloud-tencent/pull/192)
 - [Feature:Support near by router.](https://github.com/Tencent/spring-cloud-tencent/pull/196)
+- [Feature: Load application.yml and application-${profile}.yml from polaris server.](https://github.com/Tencent/spring-cloud-tencent/pull/199)
+- [feat:add rate limit of unirate.](https://github.com/Tencent/spring-cloud-tencent/pull/197)